+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Fixed
+- erroneous decoding of 32-bit signed integers as signed when they're less than 4 bytes in length
+
 ## [1.16.0] - 2021-05-13
 ### Added
 - OTP 24 to CI targets